reafactor: memindahkan pemanggilan api register dari formRegister ke Register page
